requirements.txt
asgiref==3.9.1
Django==4.2.15
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
pillow==11.3.0
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
